Booking screens refactoring completed
