+++ src/screens/DestinationCountry.js
@@ -1,50 +1,26 @@
-import React, {useState} from 'react';
+import React from 'react';
 import {KeyboardAvoidingView, ScrollView} from 'react-native';
 import DataFlight from '../components/DataFlight';
 import GoBackScreen from '../components/GoBackScreen';
-import LocationItem from '../components/LocationItem';
-import PrimaryButton from '../components/PrimaryButton';
+import SendLocation from '../components/SendLocation';
 import {styles} from '../styles/styleDestinationCountry';
 
 const DestinationCountry = ({navigation}) => {
-  const [country, setCountry] = useState('');
-  const [disabled, setDisabled] = useState(false);
-
-  const changeCountry = text => {
-    const isDisabled = text.length > 0 ? true : false;
-    setCountry(text);
-    setDisabled(isDisabled);
-  };
-
-  const goBackScreen = () => navigation.goBack();
-
-  const changeView = () => {
-    //Guardar 'country' en store para seguir con la siguiente vista.
-    console.log(country);
-    navigation.navigate('SelectDate');
-  };
-
   return (
     <ScrollView style={styles.backColor}>
       <KeyboardAvoidingView
         behavior={'height'}
         keyboardVerticalOffset={20}
         enabled={false}>
-        <GoBackScreen onPress={goBackScreen} />
+        <GoBackScreen navigation={navigation} />
         <DataFlight
           origin={'Guadalajara, México'}
           styleViewTop={styles.styleViewTop}
         />
-        <LocationItem
-          styleView={styles.location}
-          onChangeText={changeCountry}
-          textQuestion="Where will you be flying to?"
-          isDisable={!disabled}
-        />
-        <PrimaryButton
-          pressFunction={changeView}
-          textButton="Next"
-          isDisable={!disabled}
+        <SendLocation
+          navigation={navigation}
+          path="SelectDate"
+          question="Where will you be flying to?"
         />
       </KeyboardAvoidingView>
     </ScrollView>
