+++ src/styles/stylePickersView.js
