--- src/styles/styleSelectDate.js
@@ -1,16 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: 'white',
-    justifyContent: 'space-between',
-    paddingBottom: 20,
-  },
-  viewContainer: {
-    position: 'absolute',
-    top: 50,
-    width: '90%',
-    alignSelf: 'center',
-  },
-});
